[ALPS03744016] battery: fix mah may overflow issue
[Detail]
fix mah may overflow issue

MTK-Commit-Id: 632390672cc3e44d58ec9e01ef4b652ec9015c95

CR-Id: ALPS03744016
Feature: Others

+++ drivers/misc/mediatek/include/mt-plat/mtk_battery.h
@@ -34,7 +34,7 @@ struct FUELGAUGE_TEMPERATURE {
 };
 
 struct FUELGAUGE_PROFILE_STRUCT {
-	unsigned short mah;
+	unsigned int mah;
 	unsigned short voltage;
 	unsigned short resistance; /* Ohm*/
 	unsigned short percentage;
